telegram_fetcher/requirements.txt
fastapi
uvicorn
telethon
python-dotenv
requests
pydantic 